pyproject.toml
[project]
name = "prob-lyap"
version = "0.0.1"
description = "Off-policy RL for learning robust stable Lyapunov functions or nonlinear systems"
readme = "README.md"
requires-python = "~=3.11"
authors = [
    { name = "X", email = "X" }
]
dependencies = [
    "flax",
    "gymnasium",
    "gymnasium-robotics",
    "stable-baselines3",
    "distrax",
    "orbax",
    "sbx-rl",
    "click",
    "scienceplots",
    "pandas"
]

[project.optional-dependencies]
dev = ["black==23.10.1", "ruff==0.1.3", "mypy==1.6.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prob_lyap"]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
ignore = ["F722"]

[tool.ruff.isort]
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder"
]

[[tool.mypy.overrides]]
module = ["scipy", "distrax", "optax", "igraph", "SERGIO.*"]
ignore_missing_imports = true
